Update circular queue
