+++ Queue/Circular-Queue/CMakeLists.txt
@@ -4,7 +4,8 @@ project(Circular-Queue VERSION 0.1.0 LANGUAGES C CXX)
 include(CTest)
 enable_testing()
 
-add_executable(Circular-Queue main.cpp)
+add_executable(Circular-Queue main.cpp
+    circular_queue.cpp)
 
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
